drf_channels_example/requirements.txt
asgi-redis==0.14.1
asgiref==0.14.0
autobahn==0.16.0
channels==0.17.2
daphne==0.15.0
Django==1.10.1
djangorestframework==3.3.2
msgpack-python==0.4.8
redis==2.10.5
six==1.10.0
Twisted==16.4.1
txaio==2.5.1
wheel==0.24.0
zope.interface==4.3.2
